Fix: DTO 프러퍼티 속성 수정

+++ src/main/java/com/kosa/realestate/realestates/model/RealEstateSaleDTO.java
@@ -11,7 +11,7 @@ public class RealEstateSaleDTO {
   
   private int salesId;  //매매아이디
   private int realEstateId; //부동산아이디
-  private int salePrice;    //거래금액
+  private double salePrice;    //거래금액
   private String contractDate;  //계약년월일
   private String registrationDate;    //등기일자
   private String buildingName;  //아파트 동
